Move logic from PageMainContent to shared/lib
