+++ src/shared/lib/UI/index.ts
@@ -1,0 +1,4 @@
+export type { UISchema } from './model/types/UISchema';
+export { useScrolling } from './lib/hooks/useScrolling/useScrolling';
+export { uiReducer /* , uiActions */ } from './model/slice/UISlice';
+// export { getScrollByPath } from './UI/model/selectors/getScrollSave';
