+++ src/widgets/PageMainContent/index.ts
@@ -1,4 +1,1 @@
 export { PageMainContent } from './components/PageMainContent';
-export type { UISchema } from './UI/model/types/UISchema';
-export { uiReducer /* , uiActions */ } from './UI/model/slice/UISlice';
-// export { getScrollByPath } from './UI/model/selectors/getScrollSave';
